refactor(analytics): rename old analytics to analyticsV1
