+++ src/AnalyticsRules.php
@@ -7,69 +7,64 @@ class AnalyticsRules implements \ArrayAccess
     const RESOURCE_PATH = '/analytics/rules';
 
     private ApiCall $apiCall;
-    private $analyticsRules = [];
 
     public function __construct(ApiCall $apiCall)
     {
         $this->apiCall = $apiCall;
     }
 
-    public function __get($ruleName)
+    /**
+     * Create multiple analytics rules
+     * 
+     * @param array $rules Array of rule objects
+     * @return array Response from the API
+     */
+    public function create(array $rules)
     {
-        if (!isset($this->analyticsRules[$ruleName])) {
-            $this->analyticsRules[$ruleName] = new AnalyticsRule($ruleName, $this->apiCall);
-        }
-        return $this->analyticsRules[$ruleName];
-    }
-
-    public function upsert($ruleName, $params)
-    {
-        return $this->apiCall->put($this->endpoint_path($ruleName), $params);
-    }
-
-    public function retrieve()
-    {
-        return $this->apiCall->get($this->endpoint_path(), []);
-    }
-
-    private function endpoint_path($operation = null)
-    {
-        return self::RESOURCE_PATH . ($operation === null ? '' : "/" . encodeURIComponent($operation));
+        return $this->apiCall->post(self::RESOURCE_PATH, $rules);
     }
 
     /**
-     * @inheritDoc
+     * Retrieve all analytics rules
+     * 
+     * @return array Response from the API
+     */
+    public function retrieve()
+    {
+        return $this->apiCall->get(self::RESOURCE_PATH, []);
+    }
+
+    /**
+     * Get a specific rule by name
+     * 
+     * @param string $ruleName
+     * @return AnalyticsRule
+     */
+    public function __get($ruleName)
+    {
+        return new AnalyticsRule($ruleName, $this->apiCall);
+    }
+
+    /**
+     * ArrayAccess implementation for backwards compatibility
      */
     public function offsetExists($offset): bool
     {
-        return isset($this->analyticsRules[$offset]);
+        return true; // Rules can be accessed by name
     }
 
-    /**
-     * @inheritDoc
-     */
     public function offsetGet($offset): AnalyticsRule
     {
-        if (!isset($this->analyticsRules[$offset])) {
-            $this->analyticsRules[$offset] = new AnalyticsRule($offset, $this->apiCall);
-        }
-
-        return $this->analyticsRules[$offset];
+        return new AnalyticsRule($offset, $this->apiCall);
     }
 
-    /**
-     * @inheritDoc
-     */
     public function offsetSet($offset, $value): void
     {
-        $this->analyticsRules[$offset] = $value;
+        // Not implemented for read-only access
     }
 
-    /**
-     * @inheritDoc
-     */
     public function offsetUnset($offset): void
     {
-        unset($this->analyticsRules[$offset]);
+        // Not implemented for read-only access
     }
-}
+}
