+++ src/Client.php
@@ -71,14 +71,14 @@ class Client
     public Presets $presets;
 
     /**
+     * @var AnalyticsV1
+     */
+    public AnalyticsV1 $analyticsV1;
+
+    /**
      * @var Analytics
      */
     public Analytics $analytics;
-
-    /**
-     * @var AnalyticsV2
-     */
-    public AnalyticsV2 $analyticsV2;
 
     /**
      * @var Stemming
@@ -123,7 +123,7 @@ class Client
         $this->multiSearch   = new MultiSearch($this->apiCall);
         $this->presets       = new Presets($this->apiCall);
         $this->analytics     = new Analytics($this->apiCall);
-        $this->analyticsV2   = new AnalyticsV2($this->apiCall);
+        $this->analyticsV1   = new AnalyticsV1($this->apiCall);
         $this->stemming     = new Stemming($this->apiCall);
         $this->conversations = new Conversations($this->apiCall);
         $this->nlSearchModels = new NLSearchModels($this->apiCall);
